Fix many compilation warnings when all source files compiled with clang (due to CPU backend makes .cu files seen as .cpp files)
Fix Algorithm bugs on CPU backend

+++ Source/Device/CPU/AlgBinarySearchCPU.h
@@ -12,7 +12,7 @@ MRAY_HYBRID MRAY_CGPU_INLINE
 size_t LowerBound(Span<const T> range, const T& value)
 {
     auto it = std::lower_bound(range.data(), range.data() + range.size(), value);
-    return std::distance(range.data(), it);
+    return size_t(std::distance(range.data(), it));
 }
 
 }
